+++ utils/validation/validation_rules/estimated_ind_cqc_filled_posts_by_job_role_validation_rules.py
@@ -76,7 +76,7 @@ class EstimatedIndCqcFilledPostsByJobRoleValidationRules:
             IndCqcColumns.national_percentage_managers_filled_posts: 0.03,  # Analysis from April 2025 showed 6%, therefore lower limit is half less.
             IndCqcColumns.national_percentage_regulated_professions_filled_posts: 0.02,  # Analysis from April 2025 showed 4%, therefore lower limit is half less.
             IndCqcColumns.national_percentage_other_filled_posts: 0.07,  # Analysis from April 2025 showed 14%, therefore lower limit is half less.
-            IndCqcColumns.difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles: 0.0,
+            IndCqcColumns.difference_estimate_filled_posts_and_from_all_job_roles: 0.0,
             MainJobRoleLabels.activites_worker: min_value,
             MainJobRoleLabels.admin_staff: min_value,
             MainJobRoleLabels.advocacy: min_value,
@@ -122,7 +122,7 @@ class EstimatedIndCqcFilledPostsByJobRoleValidationRules:
             IndCqcColumns.national_percentage_managers_filled_posts: 0.10,  # Analysis from April 2025 showed 6%, therefore lower limit is half more.
             IndCqcColumns.national_percentage_regulated_professions_filled_posts: 0.06,  # Analysis from April 2025 showed 4%, therefore lower limit is half more.
             IndCqcColumns.national_percentage_other_filled_posts: 0.21,  # Analysis from April 2025 showed 14%, therefore lower limit is half more.
-            IndCqcColumns.difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles: 1.0,
+            IndCqcColumns.difference_estimate_filled_posts_and_from_all_job_roles: 1.0,
             MainJobRoleLabels.activites_worker: max_value,
             MainJobRoleLabels.admin_staff: max_value,
             MainJobRoleLabels.advocacy: max_value,